y2023/src/test/kotlin/adventofcode/y2023/Day01Test.kt
package adventofcode.y2023

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

internal class Day01Test {

    @Test
    fun solvePartOne() {
        val input = """
            1abc2
            pqr3stu8vwx
            a1b2c3d4e5f
            treb7uchet
            """.trimIndent()


        assertEquals(142, Day01.solvePartOne(input))
    }

    @Test
    fun solvePartTwo() {
        val input = """
            two1nine
            eightwothree
            abcone2threexyz
            xtwone3four
            4nineeightseven2
            zoneight234
            7pqrstsixteen
            """.trimIndent()

        assertEquals(281, Day01.solvePartTwo(input))
    }

}